refactor: remove descriptionPost method from models, dtos, repostiory and inclements propriety description in methods where includes post
