--- src/main/java/com/techdecode/blog/dtos/DescriptionPostDto.java
@@ -1,6 +1,0 @@
-package com.techdecode.blog.dtos;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record DescriptionPostDto(@NotBlank String type, @NotBlank String description, @NotBlank PostDto post) {
-}
