+++ src/main/java/com/techdecode/blog/dtos/PostDto.java
@@ -4,5 +4,5 @@ import jakarta.validation.constraints.NotBlank;
 
 import java.util.List;
 
-public record PostDto(@NotBlank String title, @NotBlank String bannerUrl, @NotBlank String font, @NotBlank String date, List<CommentDto> comments, List<DescriptionPostDto> descriptionPost) {
+public record PostDto(@NotBlank String title, @NotBlank String bannerUrl, @NotBlank String font, @NotBlank String date, @NotBlank String description, List<CommentDto> comments) {
 }
