--- src/main/java/com/techdecode/blog/models/DescriptionPostModel.java
@@ -1,54 +1,0 @@
-package com.techdecode.blog.models;
-
-import jakarta.persistence.*;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-@Entity
-@Table(name = "description_post")
-public class DescriptionPostModel implements Serializable {
-    private static final long serialVersionUID = 1L;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private UUID id;
-    private String type;
-    private String description;
-
-    @ManyToOne
-    @JoinColumn(name = "post_id")
-    private PostModel post;
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public PostModel getPost() {
-        return post;
-    }
-
-    public void setPost(PostModel post) {
-        this.post = post;
-    }
-}
