--- src/main/java/com/techdecode/blog/repository/DescriptionPostRepository.java
@@ -1,11 +1,0 @@
-package com.techdecode.blog.repository;
-
-import com.techdecode.blog.models.DescriptionPostModel;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.UUID;
-
-@Repository
-public interface DescriptionPostRepository extends JpaRepository<DescriptionPostModel, UUID> {
-}
